Fix Linux support
