+++ Plugins/Linux/StandaloneFileBrowser/.gitignore
@@ -1,2 +1,3 @@
 .idea
 cmake-build-debug
+build
